server/requirements.txt
requests
Django==3.1.3
Pillow==8.0.1
jinja2==3.0
gunicorn==20.1.0
contextvars
typing-extensions==4.1.1
aiohttp==3.8.1
click==8.0.4
wheel==0.37.1
multidict==4.5
ibm-cloud-sdk-core==3.10.0
ibm-watson==5.2.2
ibmcloudant==0.0.34
python-dotenv==0.21.0